requirements.txt
flask
numpy
opencv-python
pillow
tensorflow-cpu
gunicorn